requirements.txt
pygame~=2.6.1
pydantic~=2.10.4
icecream~=2.1.3
pillow~=11.1.0
